requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
groq==0.9.0
supabase==2.5.0
chromadb==0.5.0
sentence-transformers==3.0.1
pyyaml==6.0.1
pydantic==2.7.4
